About Page and .gitignore Update
Added a file type to the .gitignore file to ignore a popular hidden file type that MacOS tends to add to its files on its own. Also add a bit more information to the about page to give more information and experiemented on how to add multiple paragraphs to the page.