Uploading Testing Code
Added the code used to test, understand, and solve the api issue where the api was only returning stores from the South Carolina area. This file contains important information through extensive research on the api and solutions to the issue. It also includes important information on the behaviour of the api.